frontend: Add ProfileInfo component for user profile updates and refactor Settings page to utilize it
